ovn-agent: Implement ovsdb post_connect
There was a post_connect() method in the OVN agent extension
interface but it was not called from the agent. The patch adds ability
to extensions to implement a post_connect method that is called after
the IDL connects to an ovsdb-server and before it starts processing
initial events.

This is useful when we need to configure various things in the OVN agent
extension that rely on the data obtained from the ovsdb-server.

Change-Id: I5e91cde8566715fcb81c75513c6eea8a9b8ada4f
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>